my-mcp-server/mcp-server-demo/pyproject.toml
[project]
name = "mcp-server-demo"
version = "0.1.0"
description = "FastMCP demo server"
requires-python = ">=3.10"
dependencies = [
    "fastmcp",
    "fastapi",
    "uvicorn[standard]",
]